Added explicit casts; fixes MSVC 64bit compilation.

+++ src/convert.cpp
@@ -1372,7 +1372,7 @@ namespace {
     bool mb2wc(UINT cp, std::string& str)
     {
         if (str.empty()) return true;
-        int len = MultiByteToWideChar(cp, 0, str.c_str(), str.size(), 0, 0);
+        int len = MultiByteToWideChar(cp, 0, str.c_str(), (int)str.size(), 0, 0);
         if (len == 0) {
 #ifdef DEBUG
             EXV_DEBUG << "mb2wc: Failed to determine required size of output buffer.\n";
@@ -1381,7 +1381,7 @@ namespace {
         }
         std::vector<std::string::value_type> out;
         out.resize(len * 2);
-        int ret = MultiByteToWideChar(cp, 0, str.c_str(), str.size(), (LPWSTR)&out[0], len * 2);
+        int ret = MultiByteToWideChar(cp, 0, str.c_str(), (int)str.size(), (LPWSTR)&out[0], len * 2);
         if (ret == 0) {
 #ifdef DEBUG
             EXV_DEBUG << "mb2wc: Failed to convert the input string to a wide character string.\n";
@@ -1401,7 +1401,7 @@ namespace {
 #endif
             return false;
         }
-        int len = WideCharToMultiByte(cp, 0, (LPCWSTR)str.data(), str.size() / 2, 0, 0, 0, 0);
+        int len = WideCharToMultiByte(cp, 0, (LPCWSTR)str.data(), (int)str.size() / 2, 0, 0, 0, 0);
         if (len == 0) {
 #ifdef DEBUG
             EXV_DEBUG << "wc2mb: Failed to determine required size of output buffer.\n";
@@ -1410,7 +1410,7 @@ namespace {
         }
         std::vector<std::string::value_type> out;
         out.resize(len);
-        int ret = WideCharToMultiByte(cp, 0, (LPCWSTR)str.data(), str.size() / 2, (LPSTR)&out[0], len, 0, 0);
+        int ret = WideCharToMultiByte(cp, 0, (LPCWSTR)str.data(), (int)str.size() / 2, (LPSTR)&out[0], len, 0, 0);
         if (ret == 0) {
 #ifdef DEBUG
             EXV_DEBUG << "wc2mb: Failed to convert the input string to a multi byte string.\n";
